refactor(Member): memberSeq 자료형 DB에 맞춰서 Integer로 변경

+++ src/main/java/com/goormTable/member/entity/Member.java
@@ -24,7 +24,7 @@ public class Member {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "member_seq", nullable = false, unique = true)
-    private Long memberSeq;
+    private Integer memberSeq;
 
     @Column(name = "admin_id")
     private String adminId;
